Code refactoring for registration
