+++ src/main/java/com/fdmgroup/backend_streamhub/authenticate/dto/RegistrationRequest.java
@@ -1,0 +1,13 @@
+package com.fdmgroup.backend_streamhub.authenticate.dto;
+
+import com.fdmgroup.backend_streamhub.authenticate.model.Account;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class RegistrationRequest extends Account {
+
+}
